backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.6.0

# Financial Data
yfinance>=0.2.25
pandas>=1.3.0,<2.0  # Compatible with pandas-market-calendars
numpy>=1.24.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# Market Calendar
pandas-market-calendars>=4.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0 